Cleans the documentation and updates it with use cases
