+++ datagrowth/docs/configuration/usage.rst.inc
@@ -10,27 +10,30 @@ using the ``register_defaults`` function
     from datagrowth.configuration import register_defaults
 
 
-    class YourAppConfig(AppConfig):
+    class FooAppConfig(AppConfig):
         name = 'your_app'
 
         def ready(self):
-            register_defaults("your_app", {
-                "your_configuration": True,
-                "very_special": False
+            register_defaults("foo_app", {
+                "foo": True,
+                "bar": False
             })
 
 
-Once defaults are specified you can create and use configurations with ``create_config``
+Once defaults are specified you can create and use configurations with ``create_config`` anywhere in your code.
 ::
-
     from datagrowth.configuration import create_config
 
     config = create_config("your_app", {
-        "very_special": "definitely!"
+        "bar": "definitely!"
     })
 
-    print(config.your_configuration)
+    print(config.foo)
     # out: True
-    print(config.very_special)
+    print(config.bar)
     # out: "definitely!"
+    print(config.baz)
+    # raises ConfigurationNotFoundError
 
+This is the simplest way to use Datagrowth configurations.
+Things get a little bit more interesting when you use Django field configurations.
