--- system/specs/analysis_features.txt
@@ -1,35 +1,0 @@
-Quickest way
-
-Support K-Cluster algorithms
-
-Allow for many (automated) Characteristics, which specify how they came into being
-Feed a Characteristic to Manifestations
-
-
-class Charateristic(object):
-    label
-    data_type (string, value, boolean, data)
-    source (auto, manual etc.)
-    string
-    value
-    boolean
-    data (do not allocate mem for the text field)
-
-
-class Culture(object):
-    collective (can be None)
-    characteristic
-    individual_hash
-
-
-Compare Characteristics cross Collectives and cross community (like fashion collective with 2nd hand collective)
-    -> single table query
-
-
-Quickly build Characteristics from a given Collective
-    -> iter all individuals with specified keys that should be characteristics
-
-
-Easily apply existing Characteristics to Individuals (when adding Individuals with a POST for instance)
-    -> query all possible charateristics within collective
-    -> add rows for each characteristic if applicable
