--- system/specs/requirements.txt
@@ -1,37 +1,0 @@
-+ Allow better batch processing (Mysql has gone away bug)
-+ Allow partial POST (Construct a resource bit by bit)
-+ Allow actions on parts of results (fetch more images)
-+ Allow data private for user
-+ Support partial results
-+ Support existing algorithms (ImageTranslations, PeopleSuggestions, CityCelebrities)
-+ Store history
-+ Support oauth mechanism
-
-+ Integrate with BeautifulSoup
-- Integrate with pattern
-+ Prepare Facebook functionality
-+ Think about different views upon the same results
-+ Prepare website scraping (president-death statistics)
-
-+ Use rest_framework 3
-+ Use Python 3
-+ Use Django 1.8 or 1.7
-+ Use JSON path
-+ Use JSON schema
-- Update to latest Django 1.8
-
-- Fork and patch JSONField or switch to PostgreSQL
-+ Make websockets optional through setup
-- Finish tests
-
-- Resource (base class for Html, Shell and Email resource. Handles waiting states (202, job on queue and email send). Use Celery retries?
-- ShellResource (send a command and optionally check file updates)
-- EmailResource (send an email and poll an imap for response)
-- EmailResourceProcessor
-
-Support K-Cluster algorithms
-Compare Characteristics cross Collectives and cross community (like fashion collective with 2nd hand collective)
-Easily apply existing Characteristics to Individuals (when adding Individuals with a POST for instance)
-Quickly build Characteristics from a given Collective
-Allow for many (automated) Characteristics, which specify how they came into being
-Feed a Characteristic to Manifestations
