--- system/specs/roadmap.txt
@@ -1,28 +1,0 @@
-MUST
-- Open the manifestation config and clean before setting it to the community, allow $ where possible
-- Create new CommunityNumericCharacteristics data structure
-- Allow ranking to take place through RankProcessor.numeric_characteristics (CommunityNumericCharacteristics)
-    (instead of RankProcessor.hooks). Pass CommunityNumericCharacteristics cache key to config
-    and ignore individuals parameter)
-- $ parameter parser (variables)
-- Contribute generators with low mem usage
-- Way to resolve spirits in a multi-inherit community
-
-GREAT
-- Filter through Wikidata by claims
-- Backlink topic research
-- Parse hooks as javascript modules
-
-SHOULD
-- Collective pagination
-- Create new style people-suggestions
-- Create new style city-celebrities
-- Update readme
-
-NICE
-- Synonym and antonym API
-- | parameter parser (processors)
-- Squash migrations
-
-DREAM
-- Category distance calculation
